Added image upload, pdf generation and download.

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -19,19 +21,18 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
-    const errors = validationResult(req);
-    if(!errors.isEmpty()){
+    console.log(image)
+    if(!image){
         return res.status(422).render('admin/edit-product', { 
             docTitle: 'Add product', 
-            path: '/admin/edit-product',
+            path: '/admin/add-product',
             editing: false,
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -39,6 +40,24 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array() 
             })
     }
+    const errors = validationResult(req);
+    if(!errors.isEmpty()){
+        return res.status(422).render('admin/edit-product', { 
+            docTitle: 'Add product', 
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+            })
+    }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title, 
@@ -52,7 +71,28 @@ exports.postAddProduct = (req, res, next) => {
         console.log('Created Product')
         res.redirect('/admin/products')
     })
-    .catch(err => console.log(err))
+    .catch(err => {
+        // return res.status(500).render('admin/edit-product', { 
+        //     docTitle: 'Add product', 
+        //     path: '/admin/add-product',
+        //     editing: false,
+        //     hasError: true,
+        //     product: {
+        //         title: title,
+        //         imageUrl: imageUrl,
+        //         price: price,
+        //         description: description
+        //     },
+        //     errorMessage: 'Database operation failed, please try again',
+        //     validationErrors: []
+        //     })
+        // REDIRECTING
+        // res.redirect('/500');
+        // NEW ERROR
+        const error = new Error();
+        error.httpStatusCode = 500;
+        return next(error);
+        })
 }
 
 exports.showProducts = (req, res, next) => {
@@ -66,8 +106,10 @@ exports.showProducts = (req, res, next) => {
         })
      })
      .catch(err => {
-         console.log(err);
-     })
+        const error = new Error();
+        error.httpStatusCode = 500;
+        return next(error);
+    })
  }
 
  
@@ -93,14 +135,16 @@ exports.editProduct = (req, res, next) => {
             })
     })
     .catch(err => {
-        console.log(err)
+        const error = new Error();
+        error.httpStatusCode = 500;
+        return next(error);
     })
  }
 
  exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
@@ -112,7 +156,6 @@ exports.editProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -127,32 +170,48 @@ exports.editProduct = (req, res, next) => {
             return res.redirect('/');
         }
         product.title = updatedTitle;
-        product.imageUrl = updatedImageUrl;
         product.price = updatedPrice;
         product.description = updatedDescription;
+        if(image){
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         return product.save()
         .then(result => {
             console.log('Product updated');
             res.redirect('/admin/products');   
         })
         .catch(err => {
-            console.log(err);
-        }) 
+            const error = new Error();
+            error.httpStatusCode = 500;
+            return next(error);
+        })
     })
     .catch(err => {
-        console.log(err);
-    }) 
+        const error = new Error();
+        error.httpStatusCode = 500;
+        return next(error);
+    })
  }
 
 
  exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId)
+    .then(product => {
+        if(!product){
+            return next(new Error('Product not found!'))
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id});
+    })
     .then(() => {
         res.redirect('/admin/products')
     })
     .catch(err => {
-        console.log(err);
+        const error = new Error();
+        error.httpStatusCode = 500;
+        return next(error);
     })
  }
